Refactor package sturcture
