+++ app/src/main/java/hr/tvz/mmisic/lektirko/BookAdapter.kt
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
 import android.widget.TextView
+import hr.tvz.mmisic.lektirko.data.db.entities.BookItem
 
 class BookAdapter(
     private val context: Context,
