--- app/src/main/java/hr/tvz/mmisic/lektirko/BookQuestion.kt
@@ -1,7 +1,0 @@
-package hr.tvz.mmisic.lektirko
-
-data class BookQuestion(
-    val id: Int,
-    val question: String,
-    val answer: String
-)
