+++ app/src/main/java/hr/tvz/mmisic/lektirko/BookQuestionAdapter.kt
@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
 import android.widget.TextView
+import hr.tvz.mmisic.lektirko.data.db.entities.BookQuestion
 
 class BookQuestionAdapter(
     private val context: Context,
