+++ app/src/main/java/hr/tvz/mmisic/lektirko/data/db/entities/BookItem.kt
@@ -1,4 +1,4 @@
-package hr.tvz.mmisic.lektirko
+package hr.tvz.mmisic.lektirko.data.db.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
